+++ link-move-csv/src/test/java/com/nhl/link/move/runtime/csv/CsvExtractorFactoryTest.java
@@ -11,6 +11,7 @@ import java.io.IOException;
 import java.util.Iterator;
 
 import static org.junit.Assert.*;
+import static org.mockito.Matchers.anyMap;
 import static org.mockito.Mockito.when;
 
 public class CsvExtractorFactoryTest extends BaseExtractorFactoryTest<StreamConnector, CsvExtractorFactory> {
@@ -47,7 +48,7 @@ public class CsvExtractorFactoryTest extends BaseExtractorFactoryTest<StreamConn
 
 	@Test
 	public void testBasicIntegration() throws IOException {
-		when(getConnectorMock().getInputStream())
+		when(getConnectorMock().getInputStream(anyMap()))
 				.thenReturn(new ByteArrayInputStream("r1c1,r1c2,r1c3\nr2c1,r2c2,r2c3".getBytes()));
 		RowReader reader = getExtractorFactory().createExtractor(getConnectorMock(), getModel()).getReader(null);
 		doCheck(reader);
@@ -55,7 +56,7 @@ public class CsvExtractorFactoryTest extends BaseExtractorFactoryTest<StreamConn
 
 	@Test
 	public void testSetDelimiter() throws IOException {
-		when(getConnectorMock().getInputStream())
+		when(getConnectorMock().getInputStream(anyMap()))
 				.thenReturn(new ByteArrayInputStream("r1c1;r1c2;r1c3\nr2c1;r2c2;r2c3".getBytes()));
 		getModel().getProperties().put(CsvExtractorFactory.DELIMITER_PROPERTY, ";");
 		RowReader reader = getExtractorFactory().createExtractor(getConnectorMock(), getModel()).getReader(null);
@@ -64,7 +65,7 @@ public class CsvExtractorFactoryTest extends BaseExtractorFactoryTest<StreamConn
 
 	@Test
 	public void testSetReadFrom() throws IOException {
-		when(getConnectorMock().getInputStream())
+		when(getConnectorMock().getInputStream(anyMap()))
 				.thenReturn(new ByteArrayInputStream("k1,k2,k3\nr1c1,r1c2,r1c3\nr2c1,r2c2,r2c3".getBytes()));
 		getModel().getProperties().put(CsvExtractorFactory.READ_FROM_PROPERTY, "2");
 		RowReader reader = getExtractorFactory().createExtractor(getConnectorMock(), getModel()).getReader(null);