+++ link-move/src/main/java/com/nhl/link/move/connect/URIConnector.java
@@ -3,6 +3,7 @@ package com.nhl.link.move.connect;
 import java.io.IOException;
 import java.io.InputStream;
 import java.net.URI;
+import java.util.Map;
 
 /**
  * @since 1.4
@@ -16,7 +17,7 @@ public class URIConnector implements StreamConnector {
 	}
 
 	@Override
-	public InputStream getInputStream() throws IOException {
+	public InputStream getInputStream(Map<String, ?> parameters) throws IOException {
 		return uri.toURL().openStream();
 	}
 }